RA-20 refactor: changed the data to be transferred and changed in RecipeState
