+++ app/src/main/java/com/example/recipesapp/ui/recipes/detail/RecipeViewModel.kt
@@ -1,17 +1,14 @@
 package com.example.recipesapp.ui.recipes.detail
 
 import androidx.lifecycle.ViewModel
-import com.example.recipesapp.R
-import com.example.recipesapp.model.Ingredient
+import com.example.recipesapp.model.Recipe
 
 class RecipeViewModel : ViewModel() {
 
     data class RecipeState(
-        val id: Int = 0,
-        val title: String = "",
-        val ingredients: List<Ingredient> = emptyList(),
-        val method: List<String> = emptyList(),
-        val imageUrl: Int = R.drawable.burger,
+        val recipe: Recipe? = null,
+        val portionCount: Int = 1,
+        val isFavorite: Boolean = false,
     )
 
 }
